requirements.txt
Django>=5.2.12
djangorestframework>=3.16.1
python-dotenv>=1.0.0
python-dateutil>=2.9.0
